Add HttpRequestInterceptor
This probably won't be pushed upstream. Tangram allows a new httpHandler to be set to the mapcontroller, but when using Xamarin contructing a viable httphandler is difficult as we can't subclass the existing HttpHandler. So instead offer these interception points, which conforms to the pattern of the map touch events - allowing us to handle certain requests externally and feed the data back into Tangram.